Create FloorSearch
Floor in a Sorted Array
Given a sorted array arr[] of size N without duplicates, and given a value x. Find the floor of x in given array. Floor of x is defined as the largest element K in arr[] such that K is smaller than or equal to x.

Try to use binary search to solve this problem.

Input
First line of input contains number of integers in array, N and element whose floor is to be searched.
Last line of input contains array elements.